+++ basicalter/slice_test.go
@@ -1,6 +1,7 @@
 package basicalter_test
 
 import (
+	"strings"
 	"testing"
 
 	"github.com/jeremmfr/go-utils/basicalter"
@@ -26,6 +27,8 @@ func TestDelEmptyStrings(t *testing.T) {
 
 	if v := basicalter.DelEmptyStrings(sliceOfString); len(v) != 2 {
 		t.Errorf("DelEmptyStrings didn't remove empty string: %v", v)
+	} else if !basiccheck.EqualStringSlice(v, []string{"foo", "bar"}) {
+		t.Errorf("DelEmptyStrings didn't remove empty string: %v", v)
 	}
 }
 
@@ -34,6 +37,20 @@ func TestDelStringInSlice(t *testing.T) {
 
 	if v := basicalter.DelStringInSlice("baz", sliceOfString); len(v) != 2 {
 		t.Errorf("DelStringInSlice didn't remove 'baz': %v", v)
+	} else if !basiccheck.EqualStringSlice(v, []string{"foo", "bar"}) {
+		t.Errorf("DelStringInSlice didn't remove 'baz': %v", v)
+	}
+}
+
+func TestFilterStringsWith(t *testing.T) {
+	sliceOfString := []string{"foo", "baz", "bar", "baz"}
+
+	if v := basicalter.FilterStringsWith(sliceOfString, func(s string) bool {
+		return strings.HasPrefix(s, "ba")
+	}); len(v) != 3 {
+		t.Errorf("FilterStringsWith didn't remove foo (without prefix 'ba'): %v", v)
+	} else if !basiccheck.EqualStringSlice(v, []string{"baz", "bar", "baz"}) {
+		t.Errorf("FilterStringsWith didn't remove foo (without prefix 'ba'): %v", v)
 	}
 }
 
@@ -68,4 +85,18 @@ func TestSortStringsByLengthDec(t *testing.T) {
 	if !basiccheck.EqualStringSlice(s, desiredSlice) {
 		t.Errorf("SortStringsByLength didn't sort slice with smaller last and lexicographic order")
 	}
+}
+
+func TestReplaceStringsWith(t *testing.T) {
+	sliceOfString := []string{"Foo", "Bar", "Baz"}
+
+	basicalter.ReplaceStringsWith(sliceOfString, strings.ToLower)
+
+	if !basiccheck.EqualStringSlice(sliceOfString, []string{"foo", "bar", "baz"}) {
+		t.Errorf("ReplaceStringsWith didn't replace all strings in slice "+
+			"with the lowercase version: %v", sliceOfString)
+	}
+
+	// test empty slice
+	basicalter.ReplaceStringsWith([]string{}, strings.ToLower)
 }